requirements.txt
# Web server
flask>=3.0.0
flask-limiter>=3.5.0
gunicorn>=21.0.0

# Scraping — fully keyless, no API accounts needed
duckduckgo-search>=6.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data
pandas>=2.0.0
numpy>=1.24.0

# NLP / Sentiment
transformers>=4.35.0
torch>=2.0.0          # CPU-only ~800MB — see note below
sentencepiece>=0.1.99

# Image processing
opencv-python-headless>=4.8.0  # headless = no GUI deps, smaller build
scikit-image>=0.21.0
Pillow>=10.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=1.0.0

# NOTE: if Zeabur build times out on torch, replace the torch line with:
#   torch==2.1.0+cpu
# and add this line above it:
#   --extra-index-url https://download.pytorch.org/whl/cpu
